REFACTOR: extract modular concerns from Binding
