+++ lib/spaces/engines/transforming/divisions/bindings/binding.rb
@@ -1,7 +1,13 @@
 require 'resolv'
+require_relative 'publishing'
+require_relative 'resolving'
+require_relative 'packing'
 
 module Divisions
   class Binding < ::Divisions::TargetingSubdivision
+    include ::Divisions::Binding::Publishing
+    include ::Divisions::Binding::Resolving
+    include ::Divisions::Binding::Packing
 
     class << self
       def features; [:type, :identifier, :target_identifier, :configuration] ;end
@@ -13,8 +19,6 @@ module Divisions
 
     def embed?; type == 'embed' ;end
 
-    def implies_packable?; !embed? ;end
-
     def runtime_binding?; identifier == 'containing' ;end
 
     def runtime_type
@@ -23,44 +27,11 @@ module Divisions
 
     def configuration; struct.configuration || derived_features[:configuration] ;end
 
-    def localized
-      empty.tap do |m|
-        m.struct = struct.without(:target).tap do |s|
-          s.identifier ||= target_identifier
-          s.target_identifier = target_identifier
-        end
-      end
-    end
-
-    def flattened
-      empty.tap do |m|
-        m.struct = struct.tap do |s|
-          s.configuration = flattened_configuration
-        end
-      end
-    end
-
-    def flattened_configuration
-      unresolved_struct.merge(target_configuration).merge(configuration)
-    end
-
-    def resolved
-      super.tap do |d|
-        d.struct.configuration = Divisions::ResolvableStruct.new(struct.configuration, self).resolved
-      end
-    end
-
-    def infix_qualifier; target_identifier ;end
-
     def target_configuration
       @target_configuration ||= blueprint.binding_target.struct
     end
 
     def keys; configuration.to_h.keys ;end
-
-    def environment_variables
-      configuration.to_h.map { |k, v| "--env=#{k}=#{v}" }.join(' ')
-    end
 
     def method_missing(m, *args, &block)
       keys&.include?(m) ? configuration[m] : super
