+++ lib/spaces/engines/transforming/divisions/bindings/publishing.rb
@@ -1,0 +1,16 @@
+module Divisions
+  class Binding < ::Divisions::TargetingSubdivision
+    module Publishing
+
+      def localized
+        empty.tap do |m|
+          m.struct = struct.without(:target).tap do |s|
+            s.identifier ||= target_identifier
+            s.target_identifier = target_identifier
+          end
+        end
+      end
+
+    end
+  end
+end
